requirements.txt
selenium >= 4.4.3
robotframework >= 5.0.1
robotframework-seleniumlibrary >= 6.0.0
webdriver_manager >= 3.8.3